% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_line.R
\name{string_line}
\alias{string_line}
\title{Plot a recreation of E.J. Marey's graphical train schedule (aka a stringline
chart).}
\usage{
string_line(route = "glq_edb_via_fkk")
}
\arguments{
\item{route}{String. The name of a csv file (without extension) in the
extdata directory}
}
\value{
A plot object. A stringline chart.
}
\description{
Stations are plotted vertically according to their relative distance. This
means the slope of each line corresponds to the speed of the train.
}
\details{
This plot requires two data tables: a timetable and the distances between
stations. This can potentially lead to redundancy (repetition of stations) so
the demonstration datasets store these two tables in a special format as csv
files. The timetable is organised with a row for each journey and a column
for each station. Relative distances are stored in the headers. This function
parses this dataset and then plots a stringline chart.
}
\examples{
string_line()
string_line("london_edinburgh_test")

}
